refactor: rename parse command to seed
