+++ apps/cli/src/use-cases/seed/utils/enum.ts
