+++ apps/cli/src/use-cases/seed/utils/isRenderPropType.ts
